requirements.txt
aiohttp==3.8.1
discord.py==2.0.1
emoji==1.6.3
python-dotenv==0.21.0
pytimeparse==1.1.8
